Move ModDataMapTypes to neoforge package
